refactor: improve SSR and config management
