+++ frontend/src/app/api/runtime-config/route.ts
@@ -1,0 +1,7 @@
+import { getRuntimeConfig } from "@/utils/runtimeConfig";
+
+export const dynamic = "force-dynamic";
+
+export async function GET() {
+  return Response.json(getRuntimeConfig());
+}
