+++ frontend/src/app/layout.tsx
@@ -1,8 +1,6 @@
 import { GlobalProvider } from "@/contexts/GlobalContext";
 import "./globals.css";
 import type { Metadata } from "next";
-import { getRuntimeConfig } from "@/utils/config";
-import { generateWifiConfig, generateWiFiQRString } from "@/utils/wifi";
 
 export const metadata: Metadata = {
   title: "UniFi Voucher Manager",
@@ -19,24 +17,10 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const config = getRuntimeConfig();
-
-  let wifiConfig = null;
-  let wifiString = null;
-
-  try {
-    wifiConfig = generateWifiConfig(config);
-    wifiString = generateWiFiQRString(wifiConfig);
-  } catch (error) {
-    console.warn("Could not generate WiFi QR configuration:", error);
-  }
-
   return (
     <html lang="en" data-scroll-behavior="smooth" suppressHydrationWarning>
       <body className="antialiased">
-        <GlobalProvider wifiConfig={wifiConfig} wifiString={wifiString}>
-          {children}
-        </GlobalProvider>
+        <GlobalProvider>{children}</GlobalProvider>
       </body>
     </html>
   );
