--- frontend/src/app/welcome/WelcomeClient.tsx
@@ -1,44 +1,0 @@
-"use client";
-
-import { api } from "@/utils/api";
-import { useCallback, useEffect, useState } from "react";
-
-type Props = {
-  ssid?: string;
-};
-
-export default function WelcomeClient({ ssid }: Props) {
-  const [visited, setVisited] = useState(false);
-
-  const rotateVoucher = useCallback(async () => {
-    try {
-      await api.createRollingVoucher();
-    } catch (error: any) {
-      // Error 403 is expected if the user already created a rolling voucher
-      if (error?.status !== 403) {
-        console.error("Failed to create rolling voucher", error);
-      }
-    }
-  }, []);
-
-  useEffect(() => {
-    if (visited) return;
-
-    rotateVoucher();
-    setVisited(true);
-  }, [rotateVoucher, visited]);
-
-  return (
-    <main className="flex-center h-screen w-full px-4">
-      <div className="w-full text-center font-bold text-4xl sm:text-5xl md:text-7xl lg:text-9xl leading-snug">
-        {ssid ? (
-          <>
-            Welcome to <span className="text-brand font-mono">{ssid}</span>!
-          </>
-        ) : (
-          "Welcome!"
-        )}
-      </div>
-    </main>
-  );
-}
