+++ frontend/src/contexts/GlobalContext.tsx
@@ -2,10 +2,16 @@
 
 import { Theme } from "@/components/utils/ThemeSwitcher";
 import { useServerEvents } from "@/hooks/useServerEvents";
-import { WifiConfig } from "@/utils/wifi";
+import { DEFAULT_RUNTIME_CONFIG, RuntimeConfig } from "@/types/config";
+import {
+  generateWifiConfig,
+  generateWiFiQRString,
+  WifiConfig,
+} from "@/utils/wifi";
 import React, { createContext, useContext, useEffect, useState } from "react";
 
 type GlobalContextType = {
+  runtimeConfig: RuntimeConfig;
   wifiConfig: WifiConfig | null;
   wifiString: string | null;
   theme: Theme;
@@ -16,15 +22,15 @@ const GlobalContext = createContext<GlobalContextType | undefined>(undefined);
 
 type GlobalProviderProps = {
   children: React.ReactNode;
-  wifiConfig: WifiConfig | null;
-  wifiString: string | null;
 };
 
-export const GlobalProvider = ({
-  children,
-  wifiConfig,
-  wifiString,
-}: GlobalProviderProps) => {
+export const GlobalProvider = ({ children }: GlobalProviderProps) => {
+  const [runtimeConfig, setRuntimeConfig] = useState<RuntimeConfig>(
+    DEFAULT_RUNTIME_CONFIG,
+  );
+  const [wifiConfig, setWifiConfig] = useState<WifiConfig | null>(null);
+  const [wifiString, setWifiString] = useState<string | null>(null);
+
   const [theme, setTheme] = useState<Theme>(() => {
     if (typeof window === "undefined") return "system";
     const stored = localStorage.getItem("theme");
@@ -35,7 +41,36 @@ export const GlobalProvider = ({
 
     return "system";
   });
+
   useServerEvents();
+
+  // Load runtime configuration once after app startup
+  useEffect(() => {
+    async function loadRuntimeConfig() {
+      try {
+        const response = await fetch("/api/runtime-config", {
+          cache: "no-store",
+        });
+
+        if (!response.ok) {
+          throw new Error(`Runtime config request failed: ${response.status}`);
+        }
+
+        const config: RuntimeConfig = await response.json();
+        setRuntimeConfig(config);
+
+        const generatedWifiConfig = generateWifiConfig(config);
+        const generatedWifiString = generateWiFiQRString(generatedWifiConfig);
+
+        setWifiConfig(generatedWifiConfig);
+        setWifiString(generatedWifiString);
+      } catch (error) {
+        console.warn("Could not load runtime configuration:", error);
+      }
+    }
+
+    loadRuntimeConfig();
+  }, []);
 
   // Apply theme when changed
   useEffect(() => {
@@ -65,6 +100,7 @@ export const GlobalProvider = ({
   return (
     <GlobalContext.Provider
       value={{
+        runtimeConfig,
         wifiConfig,
         wifiString,
         theme,
