+++ frontend/src/utils/runtimeConfig.ts
@@ -1,6 +1,10 @@
+import "server-only";
+
 import { DEFAULT_RUNTIME_CONFIG, RuntimeConfig } from "@/types/config";
 import fs from "fs";
 import path from "path";
+
+let cachedConfig: RuntimeConfig | null = null;
 
 function withDefaults(config: Partial<RuntimeConfig>): RuntimeConfig {
   return {
@@ -14,6 +18,10 @@ function withDefaults(config: Partial<RuntimeConfig>): RuntimeConfig {
 }
 
 export function getRuntimeConfig(): RuntimeConfig {
+  if (cachedConfig) {
+    return cachedConfig;
+  }
+
   const file = path.join(process.cwd(), "public", "runtime-config.json");
 
   try {
